requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
gitdb==4.0.11
GitPython==3.1.41
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==69.0.3
smmap==5.0.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
